upgrading, moving away from all in-memory example
