--- src/main/java/com/accenture/cloudnative/reference/reactoropenapi/audit/Auditor.java
@@ -1,28 +1,0 @@
-package com.accenture.cloudnative.reference.reactoropenapi.audit;
-
-import com.accenture.cloudnative.reference.reactoropenapi.event.DomainEvent;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
-
-/**
- * just here to show multiple folks can subscribe
- * to the application events
- */
-@Slf4j
-@Component
-public class Auditor {
-
-  /**
-   * catches a entity change event and processes it
-   *
-   * @param de the actual domain event
-   */
-  @Async
-  @EventListener
-  public void handleEntityChange(DomainEvent de) {
-    log.info("AUDITING CHANGE: {}", de);
-  }
-
-}
