--- src/main/java/com/accenture/cloudnative/reference/reactoropenapi/domain/DomainObject.java
@@ -1,4 +1,0 @@
-package com.accenture.cloudnative.reference.reactoropenapi.domain;
-
-public interface DomainObject {
-}
