--- src/main/java/com/accenture/cloudnative/reference/reactoropenapi/event/DomainOperation.java
@@ -1,10 +1,0 @@
-package com.accenture.cloudnative.reference.reactoropenapi.event;
-
-public enum DomainOperation {
-
-  UPDATE,
-  INSERT,
-  DELETE,
-  READ
-
-}
