+++ src/main/java/com/wickedagile/apis/reference/reactoropenapi/AppConfiguration.java
@@ -1,0 +1,10 @@
+package com.wickedagile.apis.reference.reactoropenapi;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+@Configuration
+@EnableAsync
+public class AppConfiguration {
+
+}
