+++ src/main/java/com/wickedagile/apis/reference/reactoropenapi/domain/DomainObject.java
@@ -1,0 +1,4 @@
+package com.wickedagile.apis.reference.reactoropenapi.domain;
+
+public interface DomainObject {
+}
