+++ src/main/java/com/wickedagile/apis/reference/reactoropenapi/event/DomainEvent.java
@@ -1,0 +1,18 @@
+package com.wickedagile.apis.reference.reactoropenapi.event;
+
+import com.wickedagile.apis.reference.reactoropenapi.domain.DomainObject;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@ToString
+public class DomainEvent {
+
+  DomainOperation operation;
+  DomainObject object;
+
+}
