+++ src/main/java/com/wickedagile/apis/reference/reactoropenapi/event/DomainOperation.java
@@ -1,0 +1,10 @@
+package com.wickedagile.apis.reference.reactoropenapi.event;
+
+public enum DomainOperation {
+
+  UPDATE,
+  INSERT,
+  DELETE,
+  READ
+
+}
