+++ src/main/resources/application.yml
@@ -12,30 +12,47 @@ springdoc:
     path: /docs
 spring:
   application:
-    name: "reactor.SpringOpenAPICRUD"
+    name: "wickedagile.reactiveOpenAPIMongo"
     processor:
       emitWaitNanos: 10
   data:
     mongodb:
       host: "localhost"
-      port: "27017"
+      port: 27017
       database: "testdb"
-  cloud:
-    stream:
-      bindings:
-        auditor-in-0:
-          group: ${spring.application.name}
-        consumer-in-0:
-          group: ${spring.application.name}
-          destination: "vendorChange"
-        producer-out-0:
-          destination: "vendorChange"
+#  cloud:
+#    stream:
+#      bindings:
+#        auditor-in-0:
+#          group: ${spring.application.name}
+#        consumer-in-0:
+#          group: ${spring.application.name}
+#          destination: "vendorChange"
+#        producer-out-0:
+#          destination: "vendorChange"
       ## lots to configure to make production
       ## ready - https://github.com/spring-cloud/spring-cloud-stream-binder-kafka
-      kafka:
-        binder:
-          brokers: localhost
-          defaultBrokerPort: 9092
-    function:
-      definition: producer;consumer
-
+#      kafka:
+#        binder:
+#          brokers: localhost
+#          defaultBrokerPort: 9092
+#    function:
+#      definition: producer;consumer
+  ## this only exists to support embedded mongo
+  ## you can remove if not using
+  mongodb:
+   embedded:
+      version: 3.2.3
+logging:
+  file:
+    name: "logs/application.log"
+  pattern:
+    console: "%d %-5level %logger : %msg%n"
+    file: "%d %-5level [%thread] %logger : %msg%n"
+  logback:
+    rollingpolicy:
+      file-name-pattern: "logs/application.%d{yyyy-MM-dd-HH-mm}.log"
+#      max-file-size: 10M
+      max-history: 2
+  level:
+    root: INFO
