--- src/main/resources/logback.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-
-  <appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-    <layout class="ch.qos.logback.classic.PatternLayout">
-      <Pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</Pattern>
-    </layout>
-  </appender>
-
-  <appender name="SERVER" class="ch.qos.logback.core.rolling.RollingFileAppender">
-    <file>logs/application.log</file>
-    <rollingPolicy class="ch.qos.logback.core.rolling.TimeBasedRollingPolicy">
-      <!-- daily rollover -->
-      <fileNamePattern>logs/application.%d{yyyy-MM-dd-HH-mm}.log</fileNamePattern>
-      <maxHistory>2</maxHistory>
-    </rollingPolicy>
-    <encoder>
-      <Pattern>%d{HH:mm:ss.SSS} [%thread] %-5level %logger{36} - %msg%n</Pattern>
-    </encoder>
-  </appender>
-
-  <root level="debug">
-    <appender-ref ref="SERVER"/>
-  </root>
-
-  <root level="info">
-    <appender-ref ref="STDOUT"/>
-  </root>
-
-</configuration>
